server/requirements.txt
# RSS Feed API 依赖
requests>=2.26.0
datasets>=2.0.0
torch>=1.9.0
sentence-transformers>=2.2.0
huggingface_hub>=0.10.0
feedparser>=6.0.0
bs4>=0.0.1
beautifulsoup4>=4.9.3
lxml>=4.6.3
html5lib>=1.1
python-dotenv>=0.19.0
# API 依赖
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0,<2.0.0
python-multipart>=0.0.5
# Machine Learning & Embeddings
transformers>=4.11.0
numpy>=1.21.0
scikit-learn>=1.0.0
# HTTP client
httpx>=0.23.0
# Utilities
tqdm>=4.62.0
pandas>=1.3.0
# Development
types-requests>=2.26.0
types-python-dateutil>=2.8.0
types-pyyaml>=6.0.0
# Testing
pytest>=6.2.5
pytest-asyncio>=0.15.1